requirements.txt
python-dotenv
numpy
flask
mysql-connector-python